Found the problem, still broken
The problem seems to be that composite seeds don't always calculate the openings and exits correctly because an internal room seed that indicates the presence of, say, a right exit might not be one of the rightmost internal seeds, and so it gets ignored. This happens because the new composite overlaps with an old one, so the new one ends up getting fewer internal seeds than its rectangular are would suggest. Trying to fix that has led to its own issue though, and now room seeds are being assigned too many openings for some reason.